requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asyncmy==0.2.10
asyncpg==0.29.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.109.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.27.0
